+++ ConstructingACar/Constructing a car.cs
@@ -1,5 +1,4 @@
-﻿using ConstructingACar.Interfaces;
-using System;
+﻿using System;
 
 namespace ConstructingACar
 {
@@ -33,241 +32,9 @@ namespace ConstructingACar
     /// 201 - 250 km/h -> 0.0030 liter/second
     /// (When the car brakes or freewheels with getting slower, there is no fuel consumption as in modern cars, when the car "powers" the engine.)
     /// For convenience the accelerations and brakings are always linear and consumption is only for the speed at the end of every second. No considering on higher consumption while accelerating within a second
-    /// In all tests only the whole car will be tested. Feel free to write your own tests for the other classes.
+    /// As third step you have to implement the logic for an "on-board computer" (a.k.a. "body computer module" (bcm) or "Bordcomputer"). The On-Board Computer should display the following informations: trip real time trip driving time total real time total driving time trip driven distance total driven distance actual speed (additional to the drivingInformationDisplay) trip average speed total average speed actual consumption by time (from the last second/method) actual consumption by distance (from the last second/method) trip average consumption by time total average consumption by time trip average consumption by distance total average consumption by distance estimated range with actual fuel level "trip" means since enginestart. The "total"-counter keep their values beyond enginestarts. The On-Board Computer should be resetable for the "trip" and for "total". The speed-average-values are calculated by driving time (km/h) and should be rounded for 1 decimal place. The actual-consumption-by-time is calculated by second and should be rounded for 5 decimal places. The actual-consumption-by-distance is calculated in liter/100 km and should be rounded for 1 decimal place. If the car does not drive, it should return NaN. The consumption-average-by-time-values are calculated by real time (liter/second) and should be rounded for 5 decimal places. The consumption-average-by-distance-values are calculated in liter/100 km and should be rounded for 1 decimal place. The driving-distance-values are calculated in km and should have at max 2 decimal places. The estimated-range should be calculated in km and base on the consumption of the last 100 seconds. When the car is built, it should be assumed that the consumption was 4.8 Liter for the last 100 seconds. Hint: Also the values actual-consumption-by-distance and consumption-average-by-distance are calculated every second! (They are NOT calculated e.g. every 1 km or 10km. This would be much more realistic but would also make this kata much more complicated!) Remember: Every call of a method (except Refuel :D) from the car correlates to 1 second. Methods from the On-Board Computer correlates to 0 seconds. Hint: The methods "EngineStart" and "EngineStop" of the DrivingProcessor do NOT start the engine, but give the event into the DrivingProcessor, that the engine has started or stopped.
     /// </summary>
-    internal class Constructing_a_car
+    internal partial class Constructing_a_car
     {
-        public class Car : ICar
-        {
-            public IFuelTankDisplay fuelTankDisplay;
-            private IEngine engine;
-            private IFuelTank fuelTank;
-            public IDrivingInformationDisplay drivingInformationDisplay;
-            private IDrivingProcessor drivingProcessor;
-
-            public bool EngineIsRunning { get => engine.IsRunning; }
-
-            public Car(double fuelLevel = 20, int maxAcceleration = 10) // car #2
-            {
-                drivingProcessor = new DrivingProcessor(maxAcceleration);
-                drivingInformationDisplay = new DrivingInformationDisplay(drivingProcessor);
-                fuelTank = new FuelTank(fuelLevel);
-                engine = new Engine(fuelTank);
-                fuelTankDisplay = new FuelTankDisplay(fuelTank);
-            }
-
-            public void EngineStart()
-            {
-                if(!EngineIsRunning && fuelTank.FillLevel >= 0.0003)
-                    engine.Start();
-            }
-
-            public void EngineStop()
-            {
-                if(EngineIsRunning)
-                    engine.Stop();
-            }
-
-            public void Refuel(double liters)
-            {
-                fuelTank.Refuel(liters);
-            }
-
-            public void RunningIdle()
-            {
-                engine.Consume(0.0003);
-            }
-
-            public void BrakeBy(int speed)
-            {
-                drivingProcessor.ReduceSpeed(speed);
-                RunningIdle();
-            }
-
-            public void Accelerate(int speed)
-            {
-                if (engine.IsRunning)
-                {
-                    drivingProcessor.IncreaseSpeedTo(speed);
-                    if (drivingProcessor.ActualSpeed >= 1 && drivingProcessor.ActualSpeed <= 60)
-                        engine.Consume(0.0020);
-                    else if (drivingProcessor.ActualSpeed >= 61 && drivingProcessor.ActualSpeed <= 100)
-                        engine.Consume(0.0014);
-                    else if (drivingProcessor.ActualSpeed >= 101 && drivingProcessor.ActualSpeed <= 140)
-                        engine.Consume(0.0020);
-                    else if (drivingProcessor.ActualSpeed >= 141 && drivingProcessor.ActualSpeed <= 200)
-                        engine.Consume(0.0025);
-                    else if(drivingProcessor.ActualSpeed >= 141 && drivingProcessor.ActualSpeed <= 250)
-                        engine.Consume(0.0030);
-                    else
-                        RunningIdle();
-                }
-                if (speed < drivingProcessor.ActualSpeed)
-                    FreeWheel();
-            }
-
-            public void FreeWheel()
-            {
-                if (drivingProcessor.ActualSpeed > 0)
-                    drivingProcessor.ReduceSpeed(1);
-                else
-                    RunningIdle();
-            }
-        }
-
-        public class DrivingInformationDisplay : IDrivingInformationDisplay // car #2
-        {
-            private IDrivingProcessor _drivingProcessor;
-            public int ActualSpeed { get => _drivingProcessor.ActualSpeed; }
-
-            public DrivingInformationDisplay(IDrivingProcessor drivingProcessor)
-            {
-                _drivingProcessor = drivingProcessor;
-            }
-        }
-
-        public class DrivingProcessor : IDrivingProcessor // car #2
-        {
-            private int _actualSpeed;
-            private int _maxSpeed = 250;
-            private int _acceleration = 10;
-            private int _maxAcc = 20;
-            private int _minAcc = 5;
-            private int _maxBraking = 10;
-
-            public int ActualSpeed { get => _actualSpeed; }
-
-            public DrivingProcessor(int acceleration)
-            {
-                if(acceleration < _minAcc)
-                    _acceleration = _minAcc;
-                else if (acceleration > _maxAcc)
-                    _acceleration = _maxAcc;
-                else
-                    _acceleration = acceleration;
-            }
-
-            public void IncreaseSpeedTo(int speed)
-            {
-                if (speed >= _maxSpeed)
-                    speed = _maxSpeed;
-
-                if (_actualSpeed < speed)
-                {
-                    if(speed >=_actualSpeed + _acceleration)
-                        _actualSpeed += _acceleration;
-                    else
-                        _actualSpeed += speed - _actualSpeed;
-                }
-            }
-
-            public void ReduceSpeed(int speed)
-            {
-                if (speed <= 0 )
-                {
-                    speed = 0;
-                }
-                if (_actualSpeed >= speed)
-                {
-                    if (speed <= _maxBraking)
-                        _actualSpeed -= speed;
-                    else
-                        _actualSpeed -= _maxBraking;
-                }
-                else
-                    _actualSpeed -= _actualSpeed;
-            }
-        }
-
-        public class Engine : IEngine
-        {
-            private bool _isRunning = false;
-            private IFuelTank _fuelTank;
-
-            public bool IsRunning { get => _isRunning; }
-
-            public Engine(IFuelTank fuelTank)
-            {
-                _fuelTank = fuelTank;
-            }
-
-            public void Consume(double liters)
-            {
-                if (_isRunning)
-                {
-                    _fuelTank.Consume(liters);
-                    if (_fuelTank.FillLevel == 0)
-                        Stop();
-                }
-            }
-
-            public void Start()
-            {
-                _isRunning = true;
-            }
-
-            public void Stop()
-            {
-                _isRunning = false;
-            }
-        }
-
-        public class FuelTank : IFuelTank
-        {
-            private double _fillLevel;
-            private double _tankCap = 60;
-            private double _defaultFill = 20;
-            private double _reserve = 5;
-
-            public double FillLevel { get => _fillLevel; }
-            public bool IsOnReserve { get => _fillLevel < _reserve; }
-            public bool IsComplete { get => _fillLevel == _tankCap; }
-
-            public FuelTank()
-            {
-                _fillLevel = _defaultFill;
-            }
-
-            public FuelTank(double fuelLevel)
-            {
-                if (fuelLevel < 0)
-                    fuelLevel = 0;
-                if (fuelLevel > _tankCap)
-                    fuelLevel = _tankCap;
-                _fillLevel = fuelLevel;
-            }
-
-            public void Consume(double liters)
-            {
-                if (_fillLevel > 0)
-                {
-                    _fillLevel -= liters;
-                    _fillLevel = Math.Round(_fillLevel, 10);
-                }
-                else
-                    _fillLevel = 0;
-            }
-
-            public void Refuel(double liters)
-            {
-                _fillLevel += liters;
-                if (_fillLevel > _tankCap)
-                    _fillLevel = _tankCap;
-            }
-        }
-
-        public class FuelTankDisplay : IFuelTankDisplay
-        {
-            private IFuelTank _fuelTank;
-
-            public double FillLevel { get => Math.Round(_fuelTank.FillLevel, 2); }
-            public bool IsOnReserve { get => _fuelTank.IsOnReserve; }
-            public bool IsComplete { get => _fuelTank.IsComplete; }
-
-            public FuelTankDisplay(IFuelTank fuelTank)
-            {
-                _fuelTank = fuelTank;
-            }
-        }
     }
 }
