+++ ConstructingACar/Interfaces/IDrivingProcessor.cs
@@ -13,5 +13,8 @@ namespace ConstructingACar.Interfaces
         void IncreaseSpeedTo(int speed);
 
         void ReduceSpeed(int speed);
+        double ActualConsumption { get; } // car #3
+        void EngineStart(); // car #3
+        void EngineStop(); // car #3
     }
 }
